refactor to address review feedback
