+++ packages/e2e/package.json
@@ -31,6 +31,7 @@
   "devDependencies": {
     "@playwright/test": "^1.50.0",
     "@types/node": "18.19.70",
+    "dotenv": "16.4.7",
     "execa": "^7.2.0",
     "node-pty": "^1.0.0",
     "strip-ansi": "^7.1.0"
