+++ packages/e2e/setup/cli.ts
@@ -239,4 +239,9 @@ export const cliFixture = envFixture.extend<{cli: CLIProcess}>({
   },
 })
 
+export interface CLIContext {
+  cli: CLIProcess
+  appDir: string
+}
+
 export {type E2EEnv}
